+++ src/ui/handler_manager.ts
@@ -138,7 +138,6 @@ export type MapControlsScenarioOptions = {
     preZoomAroundLoc: LngLat;
     combinedEventsInProgress: EventsInProgress;
     panDelta?: Point;
-    fixedBearing?: boolean;
 };
 
 function hasChange(result: HandlerResult) {
@@ -581,7 +580,6 @@ export class HandlerManager {
             preZoomAroundLoc,
             combinedEventsInProgress,
             panDelta,
-            fixedBearing: this._map.dragPan.fixedBearing,
         });
 
         this._camera.applyUpdatedTransform(tr);
@@ -598,15 +596,14 @@ export class HandlerManager {
         deltasForHelper,
         preZoomAroundLoc,
         combinedEventsInProgress,
-        panDelta,
-        fixedBearing}: MapControlsScenarioOptions): void {
+        panDelta}: MapControlsScenarioOptions): void {
 
         const cameraHelper = this._camera.cameraHelper;
 
         cameraHelper.handleMapControlsRollPitchBearingZoom(deltasForHelper, tr);
 
         if (!terrain) {
-            cameraHelper.handleMapControlsPan(deltasForHelper, tr, preZoomAroundLoc, fixedBearing);
+            cameraHelper.handleMapControlsPan(deltasForHelper, tr, preZoomAroundLoc);
             return;
         }
 
@@ -615,14 +612,14 @@ export class HandlerManager {
                 this._terrainMovement = true;
                 this._camera.elevationFreeze = true;
             }
-            cameraHelper.handleMapControlsPan(deltasForHelper, tr, preZoomAroundLoc, fixedBearing);
+            cameraHelper.handleMapControlsPan(deltasForHelper, tr, preZoomAroundLoc);
             return;
         }
 
         if (!this._terrainMovement && (combinedEventsInProgress.drag || combinedEventsInProgress.zoom)) {
             this._terrainMovement = true;
             this._camera.elevationFreeze = true;
-            cameraHelper.handleMapControlsPan(deltasForHelper, tr, preZoomAroundLoc, fixedBearing);
+            cameraHelper.handleMapControlsPan(deltasForHelper, tr, preZoomAroundLoc);
             return;
         }
 
@@ -631,7 +628,7 @@ export class HandlerManager {
             return;
         }
 
-        cameraHelper.handleMapControlsPan(deltasForHelper, tr, preZoomAroundLoc, fixedBearing);
+        cameraHelper.handleMapControlsPan(deltasForHelper, tr, preZoomAroundLoc);
     }
 
     _fireEvents(newEventsInProgress: EventsInProgress, deactivatedHandlers: {[handlerName: string]: Event}, allowEndAnimation: boolean): void {